refactor(ark): remove dead background-refresh machinery
Removes the silent-push / scheduled-wake refresh apparatus that was
disarmed in v0.1.1 (manifest declarations dropped, custodial sweep
kill-switched). Refresh now only runs with the app in the foreground:
the sync-tick urgency sweep, LN-receive arrival coalescing, and the
notification-tap flow, all unchanged.

Deleted:
- Android: ArkBackgroundSchedulerModule/Package, ArkRefreshAlarmReceiver,
  ArkRefreshBootCompletedReceiver, ArkRefreshHeadlessTaskService,
  MainApplication registration (all inert since v0.1.1)
- iOS: ArkBackgroundScheduler native (both copies), AppDelegate silent-push
  intercept + BGTask registration, ark.refresh task identifier, processing
  background mode
- JS: scheduler.ts wake handlers, emergencyCustodialSweep.ts (never armed),
  batteryGuidance.ts (orphaned), relay ark-refresh subscribe/expiry loop
- The background-readable seed copy in Keychain: no headless wakes means
  no reason for any seed to exist outside the biometry-locked primary
  entry. A one-time boot cleanup deletes the legacy entry on installs
  that had the feature enabled.

Kept intact: the five scheduled expiry warnings (backgroundNotifications.ts
untouched), the notification-tap deep-link (extracted to
notificationHandler.ts), the reminders toggle in all four screens (now
flips the flag + requests permission only, no seed read), CoinOS payment
push registration.

tsc error set is byte-identical before/after (no new errors).
