+++ android/app/src/main/AndroidManifest.xml
@@ -18,15 +18,9 @@
     <uses-permission android:name="android.permission.USE_BIOMETRIC" />
     <uses-permission android:name="android.permission.USE_FINGERPRINT" />
     <uses-permission android:name="android.permission.POST_NOTIFICATIONS" />
-    <!-- Auto-refresh while-app-closed (FGS_DATA_SYNC) was dropped in v0.1.1:
-         the 5 scheduled local-notification warnings (4d/2d/24h/12h/6h before
-         expiry) survive without any foreground service, and Play submission
-         for FGS_DATA_SYNC requires a manual-review video that we don't have
-         enough demo value to justify. The Java/JS scheduler code is left in
-         place but is inert without the service+receiver declarations below.
-         RECEIVE_BOOT_COMPLETED is kept because react-native-push-notification
-         (RNPushNotificationBootEventReceiver) uses it to re-arm scheduled
-         local notifications after reboot. -->
+    <!-- RECEIVE_BOOT_COMPLETED is kept because react-native-push-notification
+         (RNPushNotificationBootEventReceiver) uses it to re-arm the scheduled
+         capsule-expiry warnings after reboot. -->
     <uses-permission android:name="android.permission.RECEIVE_BOOT_COMPLETED" />
     <uses-permission android:name="android.permission.VIBRATE" />
     <uses-permission android:name="android.permission.ACTION_OPEN_DOCUMENT" />
@@ -86,14 +80,9 @@
             </intent-filter>
         </service>
 
-        <!-- Ark background-refresh services and receivers (FGS_DATA_SYNC)
-             were removed in v0.1.1. The Java classes
-             (ArkRefreshHeadlessTaskService, ArkRefreshAlarmReceiver,
-             ArkRefreshBootCompletedReceiver, ArkBackgroundSchedulerModule)
-             remain in the source tree but are inert because they are no
-             longer declared here. The 5 capsule-expiry warning
-             notifications fire via PushNotification.localNotificationSchedule
-             which is independent of these services. -->
+        <!-- The 5 capsule-expiry warning notifications fire via
+             PushNotification.localNotificationSchedule and need no
+             services or receivers beyond RNPushNotification's own. -->
 
         <!-- WorkManager auto-initialization is still disabled. We don't
              currently use WM for Ark refresh, but other libraries may
