--- android/app/src/main/java/io/bluewallet/bluewallet/ArkBackgroundSchedulerPackage.java
@@ -1,34 +1,0 @@
-package io.cypherbox.btc;
-
-import androidx.annotation.NonNull;
-
-import com.facebook.react.ReactPackage;
-import com.facebook.react.bridge.NativeModule;
-import com.facebook.react.bridge.ReactApplicationContext;
-import com.facebook.react.uimanager.ViewManager;
-
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-
-/**
- * RN package wrapper for ArkBackgroundSchedulerModule. Registered from
- * MainApplication.java so the module is exposed to JS as
- * NativeModules.ArkBackgroundScheduler.
- */
-public class ArkBackgroundSchedulerPackage implements ReactPackage {
-
-    @NonNull
-    @Override
-    public List<NativeModule> createNativeModules(@NonNull ReactApplicationContext reactContext) {
-        List<NativeModule> modules = new ArrayList<>();
-        modules.add(new ArkBackgroundSchedulerModule(reactContext));
-        return modules;
-    }
-
-    @NonNull
-    @Override
-    public List<ViewManager> createViewManagers(@NonNull ReactApplicationContext reactContext) {
-        return Collections.emptyList();
-    }
-}
