+++ android/app/src/main/java/io/bluewallet/bluewallet/MainApplication.java
@@ -31,7 +31,6 @@ public class MainApplication extends Application implements ReactApplication, Co
           List<ReactPackage> packages = new PackageList(this).getPackages();
           // Hand-registered: not auto-linked because the package lives in
           // app/src/main/java rather than a published RN module.
-          packages.add(new ArkBackgroundSchedulerPackage());
           packages.add(new ArkSafBackupPackage());
           return packages;
         }
