--- src/services/ark/scheduler.ts
@@ -1,330 +1,0 @@
-import { AppRegistry, NativeEventEmitter, NativeModules, Platform } from 'react-native';
-import PushNotification from 'react-native-push-notification';
-
-import useAuthStore from '@Cypher/stores/authStore';
-
-import { navigationRef } from '../../../NavigationService';
-import { isArkExpiryWarningSource } from './backgroundNotifications';
-
-/**
- * JS-side wrapper around the native ArkBackgroundScheduler module.
- *
- * Exports a flat surface (schedule / cancel) so the rest of the codebase
- * doesn't need to care which OS-level primitive is doing the work
- * (BGAppRefreshTask on iOS, WorkManager periodic on Android).
- *
- * Also owns the registration of the platform handlers — the Android
- * HeadlessJsTask receiver and the iOS event-emitter listener — both of
- * which have to be in place at JS bundle load time, BEFORE the first
- * wake fires. See registerArkBackgroundRefreshHandlers.
- */
-
-type SchedulerNative = {
-    schedule: () => Promise<void>;
-    cancel: () => Promise<void>;
-    markTaskCompleted: (taskId: string, success: boolean) => void;
-    // Android-only — return false on iOS via the missing-method guard.
-    isIgnoringBatteryOptimizations?: () => Promise<boolean>;
-    openBatteryOptimizationSettings?: () => Promise<void>;
-    getDeviceManufacturer?: () => Promise<string>;
-    // DEMO-ONLY: synthesise the alarm-fire broadcast from inside the app's
-    // UID for the Play Console FGS_DATA_SYNC video. Remove before next
-    // release. See ArkBackgroundSchedulerModule.debugFireRefresh.
-    debugFireRefresh?: () => Promise<void>;
-};
-
-/**
- * DEMO-ONLY: trigger the AlarmManager -> Receiver -> startForegroundService
- * path on demand so the persistent FGS notification appears for the Play
- * Console submission video. Returns false if the native bridge is
- * unavailable (iOS, simulator, JS-only test runs).
- */
-export async function debugFireRefresh(): Promise<boolean> {
-    const native = getNative();
-    if (!native?.debugFireRefresh) return false;
-    await native.debugFireRefresh();
-    return true;
-}
-
-function getNative(): SchedulerNative | null {
-    return (NativeModules.ArkBackgroundScheduler as SchedulerNative | undefined) ?? null;
-}
-
-export async function scheduleArkBackgroundRefresh(): Promise<void> {
-    const native = getNative();
-    if (!native) {
-        console.warn('[Ark scheduler] native module unavailable — schedule() is a no-op');
-        return;
-    }
-    await native.schedule();
-}
-
-export async function cancelArkBackgroundRefresh(): Promise<void> {
-    const native = getNative();
-    if (!native) return;
-    await native.cancel();
-}
-
-/**
- * Android-only probe: is the OS treating Cypher Box as exempt from
- * battery-optimisation throttling?
- *
- * Returns true on iOS (no equivalent constraint — iOS background tasks
- * are managed by BGTaskScheduler with budget per app, not by a per-app
- * allowlist) and on Android <23 (pre-Doze).
- *
- * On Android 6+ (Doze era) returns the actual OS allowlist state. False
- * is the bad case: AlarmManager fires can be deferred indefinitely
- * during deep Doze, especially on Samsung One UI / Xiaomi MIUI / similar
- * vendor-specific battery managers.
- */
-export async function isIgnoringBatteryOptimizations(): Promise<boolean> {
-    if (Platform.OS !== 'android') return true;
-    const native = getNative();
-    if (!native?.isIgnoringBatteryOptimizations) return true;
-    try {
-        return await native.isIgnoringBatteryOptimizations();
-    } catch (err: any) {
-        if (__DEV__) {
-            console.log('[Ark scheduler] battery probe failed:', err?.message ?? err);
-        }
-        // Conservative: if we can't tell, assume worst-case (not ignoring)
-        // so the onboarding nudge fires instead of being silently
-        // suppressed.
-        return false;
-    }
-}
-
-/**
- * Android-only: open the system Settings page where the user can move
- * Cypher Box onto the battery-optimisation allowlist. Resolves once the
- * Settings activity has been launched (not when the user has finished
- * tapping through it — the result has to be re-probed via
- * `isIgnoringBatteryOptimizations` after the user returns).
- *
- * Falls back to generic Settings.ACTION_SETTINGS if the device's stock
- * battery-allowlist activity isn't recognised. Custom ROMs occasionally
- * route the action elsewhere; the fallback is so the tap doesn't
- * become a no-op.
- *
- * No-op on iOS.
- */
-export async function openBatteryOptimizationSettings(): Promise<void> {
-    if (Platform.OS !== 'android') return;
-    const native = getNative();
-    if (!native?.openBatteryOptimizationSettings) return;
-    await native.openBatteryOptimizationSettings();
-}
-
-/**
- * Android-only: returns the lower-cased device manufacturer string
- * (e.g. "samsung", "xiaomi", "huawei", "oneplus"). Used to pick
- * vendor-specific guidance copy in the battery onboarding modal —
- * Samsung's "Battery → Unrestricted" flow has different button labels
- * than Xiaomi's "Auto-start" toggle, etc.
- *
- * Returns empty string on iOS or on probe failure (caller treats empty
- * as "generic Android" guidance).
- */
-export async function getDeviceManufacturer(): Promise<string> {
-    if (Platform.OS !== 'android') return '';
-    const native = getNative();
-    if (!native?.getDeviceManufacturer) return '';
-    try {
-        return await native.getDeviceManufacturer();
-    } catch {
-        return '';
-    }
-}
-
-/**
- * Deep-link the user to the Ark Capsules tab and signal it to auto-refresh
- * on mount. Called from the notification tap handler — both for live taps
- * (app already running) and cold-start taps. The cold-start case is
- * delivered through the same `onNotification` callback because
- * react-native-push-notification's `configure()` default
- * `popInitialNotification: true` auto-pops the initial notification once
- * the JS bridge is ready.
- *
- * On cold-start the JS bundle is loaded but NavigationContainer may not
- * have mounted yet. We poll `navigationRef.isReady()` for up to
- * `NAV_WAIT_MS` before giving up. If the timeout fires the zustand flag
- * stays set, so an eventual manual navigation to Capsules will still
- * consume it and trigger refresh — the user just doesn't get the
- * automatic routing on this launch.
- */
-const NAV_WAIT_MS = 8000;
-
-async function dispatchArkExpiryTapRefresh(): Promise<void> {
-    try {
-        useAuthStore.getState().setArkPendingTapRefresh(true);
-    } catch (err) {
-        console.warn('[Ark scheduler] could not set tap-refresh flag:', err);
-    }
-
-    const wallet = useAuthStore.getState().arkWallet;
-    if (!wallet) {
-        // No Ark wallet on this install — the notification shouldn't have
-        // fired in the first place. Bail rather than navigate to a screen
-        // whose other tabs would dereference a null wallet.
-        return;
-    }
-
-    const start = Date.now();
-    while (!navigationRef.isReady() && Date.now() - start < NAV_WAIT_MS) {
-        await new Promise<void>(resolve => setTimeout(resolve, 100));
-    }
-    if (!navigationRef.isReady()) {
-        return;
-    }
-
-    navigationRef.current?.navigate('CheckingAccountNew', {
-        wallet,
-        accountType: 'ark',
-        initialTab: 0,
-    });
-}
-
-/**
- * One-time registration of platform task handlers. Must run from
- * index.js so both Android's AppRegistry-based headless task lookup
- * and iOS's BGTask event listener are attached before the OS routes
- * its first scheduled wake into our process.
- *
- * The orchestrator is loaded lazily inside the handler (require, not
- * import-at-top) so the index.js boot path doesn't pay the cost of
- * importing the full Ark service chain on every app start. Users who
- * never enable the toggle never load the orchestrator.
- */
-export function registerArkBackgroundRefreshHandlers(): void {
-    // Single PushNotification.configure call for the whole app. Set up
-    // here at boot so the FCM-arrival path on Android (silent push wake)
-    // has a registered onNotification before any message can arrive.
-    // iOS silent-push wake is intercepted in AppDelegate.mm via the
-    // ArkBackgroundScheduler native handler — the JS-side onNotification
-    // never sees it, so this configure is effectively Android-only for
-    // routing purposes.
-    //
-    // requestPermissions: false — opt-in flow requests permission via
-    // ensureBgNotificationPermission() when the user enables the toggle.
-    PushNotification.configure({
-        requestPermissions: false,
-        onNotification: (notification: any) => {
-            const data = (notification && (notification.data || notification.userInfo)) || {};
-            if (data && data.type === 'ark.refresh.due') {
-                if (Platform.OS === 'android') {
-                    // FCM data-message arrival has spun up the JS bridge.
-                    // Run the orchestrator inline — no WorkManager
-                    // round-trip needed since we're already alive. The
-                    // orchestrator records its own telemetry; nothing to
-                    // signal back to native.
-                    require('./backgroundRefresh')
-                        .runBackgroundRefresh('push')
-                        .then(() => {
-                            // Safety net: after refresh resolves, fire the
-                            // custodial-sweep check. Per Feature spec B,
-                            // refresh wins when it can; this only acts on
-                            // VTXOs still < CUSTODIAL_SWEEP_THRESHOLD blocks
-                            // from expiry. Idempotent + rate-limited
-                            // internally.
-                            return require('./emergencyCustodialSweep')
-                                .runEmergencyCustodialSweep('push')
-                                .catch((sweepErr: unknown) => {
-                                    console.warn('[Ark scheduler] post-push sweep threw:', sweepErr);
-                                });
-                        })
-                        .catch((err: unknown) => {
-                            console.warn('[Ark scheduler] Android push handler threw:', err);
-                        });
-                }
-                // iOS: native AppDelegate already handled it. We won't
-                // see the notification at this layer for content-available
-                // pushes — included here as defensive code in case the
-                // routing changes.
-            } else if (
-                notification &&
-                notification.userInteraction === true &&
-                isArkExpiryWarningSource(data.source)
-            ) {
-                // User tapped a pre-scheduled VTXO expiry warning (or one
-                // of the live equivalents fired from a bg-refresh tick).
-                // The earlier design ran `runBackgroundRefresh` inside
-                // this handler, which had no UI signal and could fail
-                // silently when the wallet handle was null at tap time.
-                //
-                // The new path is deterministic: set a one-shot zustand
-                // flag, then deep-link to the Capsules tab. ArkCapsules
-                // consumes the flag on mount, hydrates the wallet
-                // (re-prompts Keychain biometric if needed), and fires
-                // `refreshIds` against every imminent VTXO with the
-                // existing "Refreshing…" indicator visible from the
-                // moment the screen renders. Works on cold/background/
-                // foreground because configure() auto-pops the initial
-                // notification on cold launches.
-                void dispatchArkExpiryTapRefresh();
-            }
-        },
-    });
-
-    if (Platform.OS === 'android') {
-        AppRegistry.registerHeadlessTask('ArkBackgroundRefresh', () => async () => {
-            const { runBackgroundRefresh } = require('./backgroundRefresh');
-            const result = await runBackgroundRefresh('scheduled');
-            // Post-refresh safety net per Feature spec B. Fire-and-forget
-            // from the worker's perspective — the sweep records its own
-            // outcome via console + push, and we don't gate the
-            // WorkManager Result on it.
-            try {
-                const { runEmergencyCustodialSweep } = require('./emergencyCustodialSweep');
-                await runEmergencyCustodialSweep('scheduled');
-            } catch (sweepErr) {
-                console.warn('[Ark scheduler] post-scheduled sweep threw:', sweepErr);
-            }
-            // Return value lands in the WorkManager Result via the bridge.
-            // We always succeed at the worker level — refresh-round outcome
-            // is recorded in our own telemetry rather than reflected as a
-            // worker retry. See ArkRefreshWorker.doWork() for the rationale.
-            return result.outcome;
-        });
-        return;
-    }
-
-    if (Platform.OS === 'ios') {
-        const native = getNative();
-        if (!native) return;
-        const emitter = new NativeEventEmitter(NativeModules.ArkBackgroundScheduler as any);
-        emitter.addListener(
-            'ArkBackgroundRefreshTaskFired',
-            async (payload: { taskId?: string; trigger?: 'scheduled' | 'push' }) => {
-                const taskId = payload?.taskId;
-                const trigger = payload?.trigger ?? 'scheduled';
-                if (!taskId) return;
-                try {
-                    const { runBackgroundRefresh } = require('./backgroundRefresh');
-                    const result = await runBackgroundRefresh(trigger);
-                    // Post-refresh safety net per Feature spec B. Fire
-                    // before marking the task complete so the iOS
-                    // BGProcessingTask window covers it; idempotent +
-                    // rate-limited, so a fast no-op when nothing's
-                    // imminent.
-                    try {
-                        const { runEmergencyCustodialSweep } = require('./emergencyCustodialSweep');
-                        await runEmergencyCustodialSweep(trigger);
-                    } catch (sweepErr) {
-                        console.warn('[Ark scheduler] iOS post-refresh sweep threw:', sweepErr);
-                    }
-                    // iOS uses success/failure as a signal to the OS
-                    // scheduler. Treat any clean exit as success — only
-                    // hard errors and missing-seed cases mark failure.
-                    // Reporting failure too liberally throttles future wakes.
-                    const completedCleanly =
-                        result.outcome !== 'error' && result.outcome !== 'no_seed';
-                    native.markTaskCompleted(taskId, completedCleanly);
-                } catch (err) {
-                    console.warn('[Ark scheduler] iOS task handler threw:', err);
-                    native.markTaskCompleted(taskId, false);
-                }
-            },
-        );
-    }
-}
