+++ src/stores/authStore.ts
@@ -233,10 +233,9 @@ export type AuthStateType = {
     setWithdrawArkThreshold: (state: any) => void;
     setReserveArkAmount: (state: number) => void;
 
-    // Background VTXO refresh (opt-in). Default off — see
-    // src/services/ark/backgroundRefresh.ts for the policy and
-    // src/services/ark/backgroundKeychain.ts for the keychain trade-off
-    // the toggle gates.
+    // Reminders + foreground auto-refresh (opt-in). Gates the five
+    // scheduled expiry warnings and the sync-tick urgency sweep — see
+    // src/services/ark/backgroundRefresh.ts for the policy.
     arkBgRefreshEnabled: boolean;
     /** Timestamp (ms) of the last successful background round. Drives 12h rate limit + UI status copy. */
     arkBgRefreshLastSuccessAt: number | null;
@@ -512,12 +511,9 @@ const createAuthStore = (
             allBTCWallets: get().allBTCWallets.filter(wallet => wallet !== 'ARK'),
             // Keep thresholds — don't reset on logout
 
-            // Background-refresh state is wallet-scoped: clear on
-            // disconnect so the next wallet doesn't inherit a previous
-            // wallet's success timestamp / failure count. The keychain
-            // entry itself is cleared by setArkBackgroundRefreshEnabled
-            // (off path) which we call separately from the disconnect
-            // flow.
+            // Refresh state is wallet-scoped: clear on disconnect so
+            // the next wallet doesn't inherit a previous wallet's
+            // success timestamp / failure count.
             arkBgRefreshEnabled: false,
             arkBgRefreshLastSuccessAt: null,
             arkBgRefreshLastAttempt: null,
